ppp: Update 2.4.8 -> 2.4.9
This is the first ppp release in a long time. Many patches
were resolved upstream:

* musl fixes were merged
* EAP patch was a backport added upstream
* cflags were fixed upstream
* CVE fix was merged upstream and a backport
* pcap header from the host was fixed upstream
* suid bits during install was removed upstream

The only patch left was the /var/ redirect for resolv.conf which no longer
applied cleanly after upstream changes. For this one the patch will
need to be rewritten (and preferably submitted upstream) by someone
who needs/uses it. It was presumbaly for RO rootfs and may be resolved
by symlinks in modern system usage anyway.

Tweak the files pulled into the pppoe package for a compatibility
symlink and module rename.

Add CC to the OEMAKE command to allow builds correctly.

[Big thanks to Alex Kanavin for a lot of the work with upstream and
pre-release testing of this]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>